GH-174 :: add async support to page wrapper handler
