+++ src/TrainingGuides.Web/Features/Commerce/EventHandlers/ProductPageWrapperHandler.cs
@@ -1,640 +1,126 @@
 using CMS;
+using CMS.Base;
 using CMS.ContentEngine;
 using CMS.Core;
 using CMS.DataEngine;
-using CMS.Membership;
 
-using Kentico.PageBuilder.Web.Mvc;
-
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-
-using TrainingGuides.ProductStock;
 using TrainingGuides.Web.Commerce.EventHandlers;
-using TrainingGuides.Web.Features.Commerce.Products.Widgets.ProductListing;
-using TrainingGuides.Web.Features.Commerce.Products.Widgets.ProductWidget;
-using TrainingGuides.Web.Features.Shared.Logging;
-using TrainingGuides.Web.Features.Shared.OptionProviders.ColorScheme;
-using TrainingGuides.Web.Features.Shared.OptionProviders.ColumnLayout;
-using TrainingGuides.Web.Features.Shared.OptionProviders.CornerStyle;
-using TrainingGuides.Web.Features.Shared.Sections.General;
-using TrainingGuides.Web.Features.Shared.Services;
 
-[assembly: RegisterModule(typeof(ProductPageWrapperHandler))]
+[assembly: RegisterModule(typeof(ProductPageWrapperModule))]
 
 namespace TrainingGuides.Web.Commerce.EventHandlers;
 
-// After we created this code sample, Xperience refresh 31.3.0 added a new way to implement and register content item event handlers,
-// with full support for async/await code and dependency injection.
-// See https://docs.kentico.com/documentation/developers-and-admins/customization/handle-global-events/reference-global-system-events#contentitemevents
-public class ProductPageWrapperHandler() : Module(MODULE_NAME)
+public class ProductPageWrapperModule : Module
 {
-    // Name for initializing the module
-    public const string MODULE_NAME = "Product page wrapper handlers";
+    public ProductPageWrapperModule() : base(nameof(ProductPageWrapperModule)) { }
 
-    // Username of an admin user
-    private const string ADMIN = "administrator";
-
-    // Path to the store section of the site
-    private const string STORE_PATH = "/Store";
-
-    // Name of the website channel
-    private const string CHANNEL_NAME = "TrainingGuidesPages";
-
-    // GUID of the website channel (Note this is the value of WebsiteChannelGUID, not ChannelGUID)
-    private const string WEB_CHANNEL_GUID = "FDBA40FE-1ECE-4821-9D57-EAA1D89E13B1";
-
-
-    // We are setting these to default! to avoid a compiler warning.
-    // We know these will be initialized in the OnInit method, instead of the typical constructor DI pattern.
-    // This is a known limitation of the Module base class and does not indicate actual null safety issues in this code.
-    private IWebPageManagerFactory webPageManagerFactory = default!;
-    private IWebPageManager webPageManager = default!;
-    private IContentItemRetrieverService contentItemRetrieverService = default!;
-    private IInfoProvider<UserInfo> userInfoProvider = default!;
-    private IInfoProvider<WebsiteChannelInfo> websiteChannelInfoProvider = default!;
-    private ILogger<ProductPageWrapperHandler> logger = default!;
-
-    // Contains initialization code that is executed when the application starts
-    protected override void OnInit(ModuleInitParameters parameters)
+    protected override void OnPreInit(ModulePreInitParameters parameters)
     {
-        base.OnInit();
+        base.OnPreInit(parameters);
 
-        webPageManagerFactory = parameters.Services.GetRequiredService<IWebPageManagerFactory>();
-        contentItemRetrieverService = parameters.Services.GetRequiredService<IContentItemRetrieverService>();
-        userInfoProvider = parameters.Services.GetRequiredService<IInfoProvider<UserInfo>>();
-        websiteChannelInfoProvider = parameters.Services.GetRequiredService<IInfoProvider<WebsiteChannelInfo>>();
-        logger = parameters.Services.GetRequiredService<ILogger<ProductPageWrapperHandler>>();
-
-        var user = userInfoProvider.Get()
-            .WhereEquals(nameof(UserInfo.UserName), ADMIN)
-            .FirstOrDefault();
-
-        var webChannel = websiteChannelInfoProvider.Get()
-            .WhereEquals(nameof(WebsiteChannelInfo.WebsiteChannelGUID), new Guid(WEB_CHANNEL_GUID))
-            .FirstOrDefault();
-
-        webPageManager = webPageManagerFactory.Create(webChannel?.WebsiteChannelID ?? 0, user?.UserID ?? 0);
-
-        // Suppress CS8622: Kentico's event system delegates have nullability attribute mismatches with our handler signatures.
-        // This is a known framework limitation and does not indicate actual null safety issues in this code.
-#pragma warning disable CS8622 // Nullability of reference types in type of parameter doesn't match the target delegate (possibly because of nullability attributes).
-        ContentItemEvents.Create.After += ContentItem_Create_After;
-        ContentItemEvents.CreateLanguageVariant.After += ContentItem_CreateLanguageVariant_After;
-        ContentItemEvents.Delete.Execute += ContentItem_Delete_Execute;
-        ContentItemEvents.Publish.Execute += ContentItem_Publish_Execute;
-        ContentItemEvents.Unpublish.Execute += ContentItem_Unpublish_Execute;
-#pragma warning restore CS8622 // Nullability of reference types in type of parameter doesn't match the target delegate (possibly because of nullability attributes).
+        parameters.Services.AddSingleton<ProductPageWrapperService>();
+        parameters.Services.AddEventHandler<AfterCreateContentItemEvent, ProductPageCreateHandler>();
+        parameters.Services.AddEventHandler<AfterCreateLanguageVariantEvent, ProductPageCreateLanguageVariantHandler>();
+        parameters.Services.AddEventHandler<AfterDeleteContentItemEvent, ProductPageDeleteHandler>();
+        parameters.Services.AddEventHandler<AfterPublishContentItemEvent, ProductPagePublishHandler>();
+        parameters.Services.AddEventHandler<AfterUnpublishContentItemEvent, ProductPageUnpublishHandler>();
     }
+}
 
-    // Handler for content item creation
-    private void ContentItem_Create_After(object sender, CreateContentItemEventArgs e)
+// Handler triggered after a new product content item is created
+internal class ProductPageCreateHandler(ProductPageWrapperService service) : IAsyncEventHandler<AfterCreateContentItemEvent>
+{
+    public async Task HandleAsync(AfterCreateContentItemEvent asyncEvent, CancellationToken cancellationToken)
     {
-        if (e.ID is null || !IsApplicableType(e.ContentTypeName))
+        var data = asyncEvent.Data;
+
+        if (data.ID is null || !service.IsApplicableType(data.ContentTypeName))
             return;
 
-        // Create a web page wrapper for the newly created product
-        if (e.GUID is not null)
+        if (data.Guid is not null)
         {
-            _ = CreatePageWrapperForProduct(e.DisplayName, e.ContentLanguageName, e.ContentLanguageID, e.ContentTypeID, (Guid)e.GUID);
+            await service.CreatePageWrapperForProduct(
+                displayName: data.DisplayName,
+                languageName: data.ContentLanguageName,
+                languageId: data.ContentLanguageID,
+                contentTypeId: data.ContentTypeID,
+                contentItemGuid: data.Guid.Value);
         }
     }
+}
 
-    // Handler for content item language variant creation
-    private void ContentItem_CreateLanguageVariant_After(object sender, CreateContentItemLanguageVariantEventArgs e)
+// Handler triggered after a new language variant of a product content item is created
+internal class ProductPageCreateLanguageVariantHandler(ProductPageWrapperService service) : IAsyncEventHandler<AfterCreateLanguageVariantEvent>
+{
+    public async Task HandleAsync(AfterCreateLanguageVariantEvent asyncEvent, CancellationToken cancellationToken)
     {
-        if (!IsApplicableType(e.ContentTypeName))
+        var data = asyncEvent.Data;
+
+        if (!service.IsApplicableType(data.ContentTypeName))
             return;
 
-        _ = EnsureProductPageWrapperForLanguage(e.DisplayName, e.ContentLanguageName, e.ContentLanguageID, e.ContentTypeID, e.Guid);
+        await service.EnsureProductPageWrapperForLanguage(
+            displayName: data.DisplayName,
+            languageName: data.ContentLanguageName,
+            languageId: data.ContentLanguageID,
+            contentTypeId: data.ContentTypeID,
+            contentItemGuid: data.Guid);
     }
+}
 
-    // Handler for content item deletion
-    private void ContentItem_Delete_Execute(object sender, DeleteContentItemEventArgs e)
+// Handler triggered after a product content item is deleted
+internal class ProductPageDeleteHandler(ProductPageWrapperService service) : IAsyncEventHandler<AfterDeleteContentItemEvent>
+{
+    public async Task HandleAsync(AfterDeleteContentItemEvent asyncEvent, CancellationToken cancellationToken)
     {
-        if (!IsApplicableType(e.ContentTypeName))
+        var data = asyncEvent.Data;
+
+        if (!service.IsApplicableType(data.ContentTypeName))
             return;
 
         // If there is no existing page in the language being deleted, we don't need to create a new one, so we will not use the ensure method.
-        var productPages = GetProductPagesForProduct(e.Guid, e.ContentLanguageName, e.ContentLanguageID);
-
-        foreach (var productPage in productPages)
-        {
-            webPageManager.Delete(
-                    new DeleteWebPageParameters(productPage.SystemFields.WebPageItemID, e.ContentLanguageName)
-                    {
-                        Permanently = true,
-                    }).GetAwaiter().GetResult();
-        }
-    }
-
-    // Handler for content item publishing
-    private void ContentItem_Publish_Execute(object sender, PublishContentItemEventArgs e)
-    {
-        if (!IsApplicableType(e.ContentTypeName))
-            return;
-
-        var productPages = EnsureProductPageWrapperForLanguage(e.DisplayName, e.ContentLanguageName, e.ContentLanguageID, e.ContentTypeID, e.Guid);
-
-        foreach (var productPage in productPages)
-        {
-            if (!webPageManager.TryPublish(productPage.SystemFields.WebPageItemID, e.ContentLanguageName).GetAwaiter().GetResult())
-            {
-                logger.LogError(EventIds.ProductWrapperPublishFailed,
-                "Publish failed for product page with ID {WebPageItemID} for product content item ID {ContentItemID} in language {LanguageName}.",
-                productPage.SystemFields.WebPageItemID,
-                e.ID,
-                e.ContentLanguageName);
-            }
-        }
-    }
-
-    // Handler for content item unpublishing
-    private void ContentItem_Unpublish_Execute(object sender, UnpublishContentItemEventArgs e)
-    {
-        if (!IsApplicableType(e.ContentTypeName))
-            return;
-
-        var productPages = GetProductPagesForProduct(e.Guid, e.ContentLanguageName, e.ContentLanguageID);
-
-        foreach (var productPage in productPages)
-        {
-            if (!webPageManager.TryUnpublish(productPage.SystemFields.WebPageItemID, e.ContentLanguageName).GetAwaiter().GetResult())
-            {
-                logger.LogError(EventIds.ProductWrapperUnpublishFailed,
-                "Unpublish failed for product page with ID {WebPageItemID} for product content item ID {ContentItemID} in language {LanguageName}.",
-                productPage.SystemFields.WebPageItemID,
-                e.ID,
-                e.ContentLanguageName);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves the content type names of all product types that should have a page wrapper.
-    /// This is determined by finding all classes that implement IProductSchema but not IProductVariantSchema, and retrieving the value of their CONTENT_TYPE_NAME constant.
-    /// </summary>
-    /// <returns>A list of content type names</returns>
-    /// <remarks>
-    /// Make sure the content type of the page wrapper does not meet the criteria defined in this method, otherwise it will cause an infinite loop of page creation.
-    /// </remarks>
-    private IEnumerable<string> GetApplicableTypeNames()
-    {
-        // We know this class is stored in the Entities project, so we can use it to access the assembly
-        var entitiesAssembly = typeof(ProductAvailableStockInfo).Assembly;
-        var types = entitiesAssembly
-            .GetTypes()
-            .Where(type =>
-                type.IsClass
-                && !type.IsAbstract
-                && typeof(IProductSchema).IsAssignableFrom(type)
-                && !typeof(IProductVariantSchema).IsAssignableFrom(type));
-
-        return types.Select(type =>
-        {
-            var field = type.GetField("CONTENT_TYPE_NAME");
-            if (field is not null)
-            {
-                // The field is a constant, so we don't need an instance to retrieve its value - we can use null instead
-                return field.GetValue(null) as string;
-            }
-            // If there is no CONTENT_TYPE_NAME constant, we skip this type
-            return null;
-        })
-        // Filter out the classes with no CONTENT_TYPE_NAME constant
-        .Where(name => name is not null)!;
-    }
-
-    /// <summary>
-    /// Determines whether the specified content type name is applicable for page wrapper creation.
-    /// </summary>
-    /// <param name="contentTypeName">Name of the content type to evaluate</param>
-    /// <returns>True if a page wrapper should be created for the specified content type</returns>
-    private bool IsApplicableType(string contentTypeName)
-    {
-        var types = GetApplicableTypeNames();
-        return types.Contains(contentTypeName);
-    }
-
-    /// <summary>
-    /// Creates Page Builder JSON for the product wrapper page, with a product widget configured to display the current product 
-    /// </summary>
-    /// <returns>JSON string representing the product page's Page Builder content</returns>
-    private string GetProductWidgetsConfiguration()
-    {
-        var config = new EditableAreasConfiguration
-        {
-            EditableAreas =
-            {
-                new EditableAreaConfiguration
-                {
-                    Identifier = "areaMain",
-                    Sections =
-                    {
-                        new SectionConfiguration
-                        {
-                            Identifier = Guid.NewGuid(),
-                            TypeIdentifier = ComponentIdentifiers.Sections.GENERAL,
-                            Properties = new GeneralSectionProperties
-                            {
-                                ColorScheme = ColorSchemeOption.TransparentDark.ToString(),
-                                CornerStyle = CornerStyleOption.Round.ToString(),
-                                ColumnLayout = ColumnLayoutOption.OneColumn.ToString(),
-                            },
-                            Zones =
-                            {
-                                new ZoneConfiguration
-                                {
-                                    Identifier = Guid.NewGuid(),
-                                    Name = "zoneMain",
-                                    Widgets =
-                                    {
-                                        new WidgetConfiguration
-                                        {
-                                            Identifier = Guid.NewGuid(),
-                                            TypeIdentifier = ComponentIdentifiers.Widgets.PRODUCT,
-                                            Variants =
-                                            {
-                                                new WidgetVariantConfiguration
-                                                {
-                                                    Identifier = Guid.NewGuid(),
-                                                    Properties = new ProductWidgetProperties
-                                                    {
-                                                        DisplayCurrentPage = true,
-                                                        ShowVariantSelection = true,
-                                                        ShowVariantDetails = true,
-                                                        ShowCallToAction = false,
-                                                        CallToActionText = "View Product",
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        return JsonConvert.SerializeObject(config, Formatting.None, GetSerializerSettings());
-    }
-
-    /// <summary>
-    /// Creates Page Builder JSON for the parent page of the product wrapper, with a product listing widget configured to display products that exist beneath it in the tree
-    /// </summary>
-    /// <returns>JSON string representing the parent page's Page Builder content</returns>
-    private string GetParentWidgetsConfiguration()
-    {
-        var config = new EditableAreasConfiguration
-        {
-            EditableAreas =
-            {
-                new EditableAreaConfiguration
-                {
-                    Identifier = "areaMain",
-                    Sections =
-                    {
-                        new SectionConfiguration
-                        {
-                            Identifier = Guid.NewGuid(),
-                            TypeIdentifier = ComponentIdentifiers.Sections.GENERAL,
-                            Properties = new GeneralSectionProperties
-                            {
-                                ColorScheme = ColorSchemeOption.TransparentDark.ToString(),
-                                CornerStyle = CornerStyleOption.Round.ToString(),
-                                ColumnLayout = ColumnLayoutOption.OneColumn.ToString(),
-                            },
-                            Zones =
-                            {
-                                new ZoneConfiguration
-                                {
-                                    Identifier = Guid.NewGuid(),
-                                    Name = "zoneMain",
-                                    Widgets =
-                                    {
-                                        new WidgetConfiguration
-                                        {
-                                            Identifier = Guid.NewGuid(),
-                                            TypeIdentifier = ComponentIdentifiers.Widgets.PRODUCT_LISTING,
-                                            Variants =
-                                            {
-                                                new WidgetVariantConfiguration
-                                                {
-                                                    Identifier = Guid.NewGuid(),
-                                                    Properties = new ProductListingWidgetProperties
-                                                    {
-                                                        SecuredItemsDisplayMode = SecuredOption.PromptForLogin.ToString(),
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        return JsonConvert.SerializeObject(config, Formatting.None, GetSerializerSettings());
-    }
-
-    /// <summary>
-    /// Creates JSON serializer settings for dealing with Page Builder JSON
-    /// </summary>
-    /// <returns>JSON serializer settings</returns>
-    private JsonSerializerSettings GetSerializerSettings() =>
-        new()
-        {
-            ContractResolver = new CamelCasePropertyNamesContractResolver(),
-            NullValueHandling = NullValueHandling.Include,
-        };
-
-    /// <summary>
-    /// Creates page template configuration JSON for the product wrapper page and its parent page, using the General template with default properties
-    /// </summary>
-    /// <returns>JSON string representing the page template configuration</returns>
-    private string GetPageTemplateConfiguration() =>
-        "{\"identifier\":\"TrainingGuides.GeneralPageTemplate\",\"properties\":null,\"fieldIdentifiers\":null}";
-
-    /// <summary>
-    /// Ensures a parent page exists for the specified product content type in the current language
-    /// </summary>
-    /// <param name="contentTypeId">The ID of the product content type</param>
-    /// <param name="languageName">The language in which to ensure the parent page exists</param>
-    /// <param name="languageId">The ID of the language</param>
-    /// <returns>The WebPageItemID of the parent page</returns>
-    private int EnsureParentPageInLanguage(int contentTypeId, string languageName, int languageId)
-    {
-        var contentType = DataClassInfoProvider.GetDataClassInfo(contentTypeId);
-        var contentTypeGuid = contentType?.ClassGUID ?? Guid.Empty;
-
-        // Get any language version of the parent page
-        var existingParentPages = contentItemRetrieverService
-            .RetrieveWebPageChildrenByPathWithoutContext(
-                contentTypeNames: [StoreSection.CONTENT_TYPE_NAME],
-                parentPagePath: STORE_PATH,
-                customContentTypeQueryParameters: query => query,
-                customContentQueryParameters: config => config
-                    // Filter store section pages that are assigned to the product's content type
-                    .Where(where => where.WhereContains(nameof(StoreSection.StoreSectionContentTypes), contentTypeGuid.ToString()))
-                    .TopN(1),
-                forPreview: true,
-                includeSecuredItems: true,
-                depth: 0,
-                languageName: null,
-                channelName: CHANNEL_NAME)
-            .GetAwaiter().GetResult();
-
-        if (!existingParentPages.Any())
-        {
-            // If there is no parent page for the product type, create one and return its ID
-            return CreateParentPage(
-                displayName: contentType?.ClassDisplayName ?? "Store section",
-                languageName: languageName,
-                contentTypeGuid: contentTypeGuid
-            );
-        }
-        else
-        {
-            // If there are no language variants in the specified language, create one
-            if (!existingParentPages
-                .Where(page => page.SystemFields.ContentItemCommonDataContentLanguageID == languageId)
-                .Any())
-            {
-                CreateParentPageLanguageVariant(
-                    displayName: contentType?.ClassDisplayName ?? "Store section",
-                    languageName: languageName,
-                    contentTypeGuid: contentTypeGuid,
-                    webPageItemID: existingParentPages.First().SystemFields.WebPageItemID);
-            }
-
-            // Return the existing parent page ID
-            return existingParentPages.First().SystemFields.WebPageItemID;
-        }
-    }
-
-    /// <summary>
-    /// Creates a parent page for a given product content type and language
-    /// </summary>
-    /// <param name="displayName">The display name of the parent page</param>
-    /// <param name="languageName">The language in which to create the parent page</param>
-    /// <param name="contentTypeGuid">The GUID of the product content type</param>
-    /// <returns>The WebPageItemID of the created parent page</returns>
-    private int CreateParentPage(string displayName, string languageName, Guid contentTypeGuid)
-    {
-        var parentData = new ContentItemData(new Dictionary<string, object>
-        {
-            { nameof(StoreSection.StoreSectionContentTypes), new List<Guid>(){ contentTypeGuid } },
-        });
-
-        var parentContentItemParameters = new ContentItemParameters(StoreSection.CONTENT_TYPE_NAME, parentData);
-
-        var createParentPageParameters = new CreateWebPageParameters(displayName,
-            languageName,
-            parentContentItemParameters)
-        {
-            ParentWebPageItemID = GetStorePageId(languageName) ?? 0
-        };
-
-        createParentPageParameters.SetPageBuilderConfiguration(GetParentWidgetsConfiguration(), GetPageTemplateConfiguration());
-
-        return webPageManager.Create(createParentPageParameters).GetAwaiter().GetResult();
-    }
-
-    /// <summary>
-    /// Creates a language variant for an existing parent page.
-    /// </summary>
-    /// <param name="displayName">The display name of the parent page</param>
-    /// <param name="languageName">The language in which to create the language variant</param>
-    /// <param name="contentTypeGuid">The GUID of the product content type</param>
-    /// <param name="webPageItemID">The WebPageItemID of the parent page</param>
-    private void CreateParentPageLanguageVariant(string displayName, string languageName, Guid contentTypeGuid, int webPageItemID)
-    {
-        var parentData = new ContentItemData(new Dictionary<string, object>
-        {
-            { nameof(StoreSection.StoreSectionContentTypes), new List<Guid>(){ contentTypeGuid } },
-        });
-        var createLanguageVariantParameters =
-            new CMS.Websites.CreateLanguageVariantParameters(webPageItemID,
-                                                             languageName,
-                                                             displayName,
-                                                             parentData);
-
-        createLanguageVariantParameters.SetPageBuilderConfiguration(GetParentWidgetsConfiguration(), GetPageTemplateConfiguration());
-
-        if (!webPageManager.TryCreateLanguageVariant(createLanguageVariantParameters).GetAwaiter().GetResult())
-        {
-            logger.LogError(EventIds.ProductParentPageLanguageVariantCreateFailed,
-                "Parent page language variant creation failed for product content type with GUID {ContentTypeGuid} in language {LanguageName}.",
-                contentTypeGuid,
-                languageName);
-        }
-    }
-
-    /// <summary>
-    /// Retrieves the page ID of the store's main page
-    /// </summary>
-    /// <param name="languageName">The language to query</param>
-    /// <returns>The WebPageItemID of the store's main page, or null if not found</returns>
-    private int? GetStorePageId(string languageName) =>
-        contentItemRetrieverService.RetrieveWebPageByPathWithoutContext<EmptyPage>(
-                pathToMatch: STORE_PATH,
-                includeSecuredItems: true,
-                languageName: languageName,
-                channelName: CHANNEL_NAME,
-                forPreview: true)
-            .GetAwaiter().GetResult()?.SystemFields.WebPageItemID;
-
-    /// <summary>
-    /// Creates a page wrapper for the specified product content item in the specified language, and returns the ID of the page. Ensures the parent page exists.
-    /// </summary>
-    /// <param name="displayName">The display name of the product content item, used for naming the page wrapper.</param>
-    /// <param name="languageName">The language for which to create the page wrapper.</param>
-    /// <param name="languageId">The ID of the language for which to create the page wrapper.</param>
-    /// <param name="contentTypeId">The content type ID of the product content item, used to find or create a parent page if necessary.</param>
-    /// <param name="contentItemGuid">The GUID of the product content item, used to link the page wrapper to the product.</param>
-    /// <returns>The ID of the created page.</returns>
-    /// </summary>
-    private int CreatePageWrapperForProduct(string displayName, string languageName, int languageId, int contentTypeId, Guid contentItemGuid)
-    {
-        var itemData = new ContentItemData(new Dictionary<string, object>
-        {
-            { nameof(ProductPage.ProductPageProducts), new List<ContentItemReference>()
-                { new() { Identifier = contentItemGuid } } },
-        });
-
-        var contentItemParameters = new ContentItemParameters(ProductPage.CONTENT_TYPE_NAME, itemData);
-
-        var createPageParameters = new CreateWebPageParameters(displayName, languageName, contentItemParameters)
-        {
-            ParentWebPageItemID = EnsureParentPageInLanguage(contentTypeId, languageName, languageId)
-        };
-
-        createPageParameters.SetPageBuilderConfiguration(GetProductWidgetsConfiguration(), GetPageTemplateConfiguration());
-        int id = webPageManager.Create(createPageParameters).GetAwaiter().GetResult();
-
-        if (id <= 0)
-        {
-            logger.LogError(EventIds.ProductWrapperCreateFailed,
-                "Page wrapper creation failed for product content item with GUID {ContentItemGuid} in language {LanguageName}.",
-                contentItemGuid,
-                languageName);
-        }
-
-        return id;
-    }
-
-    /// <summary>
-    /// Creates a language variant of a product's page wrapper in the specified language
-    /// </summary>
-    /// <param name="displayName">The display name of the content item to apply to the page wrapper</param>
-    /// <param name="languageName">The language in which to create the variant</param>
-    /// <param name="contentItemGuid">The GUID of the content item to be referenced by the page wrapper</param>
-    /// <param name="existingPageId">The WebPageItemID of the existing page</param>
-    private void CreatePageWrapperLanguageVariantForProduct(string displayName, string languageName, Guid contentItemGuid, int existingPageId)
-    {
-        var itemData = new ContentItemData(new Dictionary<string, object>
-        {
-            { nameof(ProductPage.ProductPageProducts), new List<ContentItemReference>()
-                { new() { Identifier = contentItemGuid } } },
-        });
-
-        var createLanguageVariantParameters =
-            new CMS.Websites.CreateLanguageVariantParameters(existingPageId,
-                                                             languageName,
-                                                             displayName,
-                                                             itemData);
-
-        createLanguageVariantParameters.SetPageBuilderConfiguration(GetProductWidgetsConfiguration(), GetPageTemplateConfiguration());
-
-        if (!webPageManager.TryCreateLanguageVariant(createLanguageVariantParameters).GetAwaiter().GetResult())
-        {
-            logger.LogError(EventIds.ProductWrapperLanguageVariantCreateFailed,
-                "Page wrapper language variant creation failed for product content item with GUID {ContentItemGuid} in language {LanguageName}.",
-                contentItemGuid,
-                languageName);
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all product pages that reference the specified product.
-    /// </summary>
-    /// <param name="guid">GUID of the product item being referenced</param>
-    /// <param name="languageName">Language version to check</param>
-    /// <param name="languageId">Language ID to check (must match the language name)</param>
-    /// <returns>A collection of product pages that reference the specified product</returns>
-    private IEnumerable<IWebPageFieldsSource> GetProductPagesForProduct(Guid guid, string? languageName, int? languageId)
-    {
-        Func<ContentQueryParameters, ContentQueryParameters> customContentQueryParameters;
-
-        Func<ContentQueryParameters, ContentQueryParameters> whereFilter = config => config
-            // Normally we would use the .Linking(...) method in the customContentTypeQueryParameters function.
-            // However, during the delete event, the linking data is already removed,
-            // so we must manually check for the GUID in ProductPageProducts.
-            .Where(where => where.WhereContains(nameof(ProductPage.ProductPageProducts), guid.ToString()));
-
-        customContentQueryParameters = languageId is int intLanguageId
-            ? (config => whereFilter(config
-                // Filter by language ID to ensure we only get pages in the specific language, without falling back to other languages
-                .Where(where => where.WhereEquals(
-                    nameof(ProductPage.SystemFields.ContentItemCommonDataContentLanguageID), intLanguageId))))
-            : whereFilter;
-
-        return contentItemRetrieverService
-                .RetrieveWebPageChildrenByPathWithoutContext(
-                    contentTypeNames: [ProductPage.CONTENT_TYPE_NAME],
-                    parentPagePath: STORE_PATH,
-                    customContentTypeQueryParameters: query => query,
-                    customContentQueryParameters: customContentQueryParameters,
-                    forPreview: true,
-                    includeSecuredItems: true,
-                    depth: 0,
-                    languageName: languageName,
-                    channelName: CHANNEL_NAME)
-                .GetAwaiter().GetResult();
-    }
-
-    /// <summary>
-    /// Ensures that a product page exists for the specified product and language. If it doesn't exist, it will be created. If a page exists in another language, a language variant will be created for the current language.
-    /// </summary>
-    /// <param name="displayName">Display name for new variant if creation required</param>
-    /// <param name="languageName">Language name of variant to retrieve</param>
-    /// <param name="languageId">Language ID of variant to retrieve (must match language name)</param>
-    /// <param name="contentTypeId">ID of the reusable product item's content type</param>
-    /// <param name="contentItemGuid">GUID of the reusable content item</param>
-    /// <returns>An enumerable collection of wrapper pages linking the specified product</returns>
-    private IEnumerable<IWebPageFieldsSource> EnsureProductPageWrapperForLanguage(string displayName, string languageName, int languageId, int contentTypeId, Guid contentItemGuid)
-    {
-        var langSpecificProductPages = GetProductPagesForProduct(contentItemGuid, languageName, languageId);
-
-        if (langSpecificProductPages.Any())
-        {
-            return langSpecificProductPages;
-        }
-
-        var allLanguageProductPages = GetProductPagesForProduct(contentItemGuid, null, null);
-
-        if (allLanguageProductPages.Any())
-        {
-            var uniquePageIDs = allLanguageProductPages.Select(page => page.SystemFields.WebPageItemID).Distinct();
-
-            foreach (int pageId in uniquePageIDs)
-            {
-                // We should still make sure the parent exists in the current language, but we don't need its ID to create a new page
-                _ = EnsureParentPageInLanguage(contentTypeId, languageName, languageId);
-
-                CreatePageWrapperLanguageVariantForProduct(displayName, languageName, contentItemGuid, pageId);
-            }
-
-            return GetProductPagesForProduct(contentItemGuid, languageName, languageId);
-        }
-
-        _ = CreatePageWrapperForProduct(displayName, languageName, languageId, contentTypeId, contentItemGuid);
-
-        return GetProductPagesForProduct(contentItemGuid, languageName, languageId);
+        await service.DeleteProductPageWrappers(
+            guid: data.Guid,
+            languageName: data.ContentLanguageName,
+            languageId: data.ContentLanguageID);
     }
 }
+
+// Handler triggered after a product content item is published
+internal class ProductPagePublishHandler(ProductPageWrapperService service) : IAsyncEventHandler<AfterPublishContentItemEvent>
+{
+    public async Task HandleAsync(AfterPublishContentItemEvent asyncEvent, CancellationToken cancellationToken)
+    {
+        var data = asyncEvent.Data;
+
+        if (!service.IsApplicableType(data.ContentTypeName))
+            return;
+
+        await service.PublishProductPageWrappers(
+            displayName: data.DisplayName,
+            languageName: data.ContentLanguageName,
+            languageId: data.ContentLanguageID,
+            contentTypeId: data.ContentTypeID,
+            contentItemGuid: data.Guid,
+            contentItemId: data.ID);
+    }
+}
+
+// Handler triggered after a product content item is unpublished
+internal class ProductPageUnpublishHandler(ProductPageWrapperService service) : IAsyncEventHandler<AfterUnpublishContentItemEvent>
+{
+    public async Task HandleAsync(AfterUnpublishContentItemEvent asyncEvent, CancellationToken cancellationToken)
+    {
+        var data = asyncEvent.Data;
+
+        if (!service.IsApplicableType(data.ContentTypeName))
+            return;
+
+        await service.UnpublishProductPageWrappers(
+            contentItemGuid: data.Guid,
+            languageName: data.ContentLanguageName,
+            languageId: data.ContentLanguageID,
+            contentItemId: data.ID);
+    }
+}
+
